+++ R/vhar-bayes.R
@@ -226,17 +226,15 @@ vhar_bayes <- function(y,
       )
     }
   } else if (prior_nm == "SSVS") {
-    # init_coef <- 1L
-    # init_coef_dummy <- 1L
-    if (length(bayes_spec$coef_spike) == 1) {
-      bayes_spec$coef_spike <- rep(bayes_spec$coef_spike, num_phi)
-    }
-    if (length(bayes_spec$coef_slab) == 1) {
-      bayes_spec$coef_slab <- rep(bayes_spec$coef_slab, num_phi)
-    }
-    if (length(bayes_spec$coef_mixture) == 1) {
-      bayes_spec$coef_mixture <- rep(bayes_spec$coef_mixture, num_grp)
-    }
+    # if (length(bayes_spec$coef_spike) == 1) {
+    #   bayes_spec$coef_spike <- rep(bayes_spec$coef_spike, num_phi)
+    # }
+    # if (length(bayes_spec$coef_slab) == 1) {
+    #   bayes_spec$coef_slab <- rep(bayes_spec$coef_slab, num_phi)
+    # }
+    # if (length(bayes_spec$coef_mixture) == 1) {
+    #   bayes_spec$coef_mixture <- rep(bayes_spec$coef_mixture, num_grp)
+    # }
     if (length(bayes_spec$coef_s1) == 2) {
       # bayes_spec$coef_s1 <- rep(bayes_spec$coef_s1, num_grp)
       coef_s1 <- numeric(num_grp)
@@ -256,40 +254,40 @@ vhar_bayes <- function(y,
     if (length(bayes_spec$rate) == 1) {
       bayes_spec$rate <- rep(bayes_spec$rate, dim_data)
     }
-    if (length(bayes_spec$chol_spike) == 1) {
-      bayes_spec$chol_spike <- rep(bayes_spec$chol_spike, num_eta)
-    }
-    if (length(bayes_spec$chol_slab) == 1) {
-      bayes_spec$chol_slab <- rep(bayes_spec$chol_slab, num_eta)
-    }
-    if (length(bayes_spec$chol_mixture) == 1) {
-      bayes_spec$chol_mixture <- rep(bayes_spec$chol_mixture, num_eta)
-    }
-    if (all(is.na(bayes_spec$coef_spike)) || all(is.na(bayes_spec$coef_slab))) {
-      # Conduct semiautomatic function using var_lm()
-      stop("Specify spike-and-slab of coefficients.")
-    }
-    if (all(is.na(bayes_spec$chol_spike)) || all(is.na(bayes_spec$chol_slab))) {
-      # Conduct semiautomatic function using var_lm()
-      stop("Specify spike-and-slab of cholesky factor.")
-    }
-    if (!(
-      length(bayes_spec$coef_spike) == num_phi &&
-        length(bayes_spec$coef_slab) == num_phi &&
-        length(bayes_spec$coef_mixture) == num_grp
-    )) {
-      stop("Invalid 'coef_spike', 'coef_slab', and 'coef_mixture' size. The vector size should be the same as 3 * dim^2.")
-    }
+    # if (length(bayes_spec$chol_spike) == 1) {
+    #   bayes_spec$chol_spike <- rep(bayes_spec$chol_spike, num_eta)
+    # }
+    # if (length(bayes_spec$chol_slab) == 1) {
+    #   bayes_spec$chol_slab <- rep(bayes_spec$chol_slab, num_eta)
+    # }
+    # if (length(bayes_spec$chol_mixture) == 1) {
+    #   bayes_spec$chol_mixture <- rep(bayes_spec$chol_mixture, num_eta)
+    # }
+    # if (all(is.na(bayes_spec$coef_spike)) || all(is.na(bayes_spec$coef_slab))) {
+    #   # Conduct semiautomatic function using var_lm()
+    #   stop("Specify spike-and-slab of coefficients.")
+    # }
+    # if (all(is.na(bayes_spec$chol_spike)) || all(is.na(bayes_spec$chol_slab))) {
+    #   # Conduct semiautomatic function using var_lm()
+    #   stop("Specify spike-and-slab of cholesky factor.")
+    # }
+    # if (!(
+    #   length(bayes_spec$coef_spike) == num_phi &&
+    #     length(bayes_spec$coef_slab) == num_phi &&
+    #     length(bayes_spec$coef_mixture) == num_grp
+    # )) {
+    #   stop("Invalid 'coef_spike', 'coef_slab', and 'coef_mixture' size. The vector size should be the same as 3 * dim^2.")
+    # }
     if (!(length(bayes_spec$shape) == dim_data && length(bayes_spec$rate) == dim_data)) {
       stop("Size of SSVS 'shape' and 'rate' vector should be the same as the time series dimension.")
     }
-    if (!(
-      length(bayes_spec$chol_spike) == num_eta &&
-        length(bayes_spec$chol_slab) == length(bayes_spec$chol_spike) &&
-        length(bayes_spec$chol_mixture) == length(bayes_spec$chol_spike)
-    )) {
-      stop("Invalid 'chol_spike', 'chol_slab', and 'chol_mixture' size. The vector size should be the same as dim * (dim - 1) / 2.")
-    }
+    # if (!(
+    #   length(bayes_spec$chol_spike) == num_eta &&
+    #     length(bayes_spec$chol_slab) == length(bayes_spec$chol_spike) &&
+    #     length(bayes_spec$chol_mixture) == length(bayes_spec$chol_spike)
+    # )) {
+    #   stop("Invalid 'chol_spike', 'chol_slab', and 'chol_mixture' size. The vector size should be the same as dim * (dim - 1) / 2.")
+    # }
     param_prior <- bayes_spec
     param_init <- lapply(
       param_init,
@@ -299,13 +297,17 @@ vhar_bayes <- function(y,
         init_coef_dummy <- rbinom(num_phi, 1, .5) # minnesota structure?
         chol_mixture <- runif(num_eta, -1, 1)
         chol_mixture <- exp(chol_mixture) / (1 + exp(chol_mixture))
-        init_chol_dummy <- rbinom(num_eta, 1, .5)
+        # init_chol_dummy <- rbinom(num_eta, 1, .5)
+        init_coef_slab <- exp(runif(num_phi, -1, 1))
+        init_contem_slab <- exp(runif(num_eta, -1, 1))
         append(
           init,
           list(
             init_coef_dummy = init_coef_dummy,
             coef_mixture = coef_mixture,
-            chol_mixture = chol_mixture
+            coef_slab = init_coef_slab,
+            chol_mixture = chol_mixture,
+            contem_slab = init_contem_slab
           )
         )
       }